Bug #33144650 FIX COMPILER WARNINGS FROM VISUAL STUDIO [group_replication, noclose]
This patch fixes, or silences, warnings so that the target
'group_replicaton' has a clean build with --config Debug.

 - Silence warnings from Visual Studio about:
     needs to have dll-interface
   when compiling code generated by protobuf.

 - Remove 'noreturn' annotation for two functions.
   Add comments that they call functions which do not return.

 - Fix some bool/int conversion warnings.

+++ plugin/group_replication/src/gcs_event_handlers.cc
@@ -1256,8 +1256,8 @@ int Plugin_gcs_events_handler::process_local_exchanged_data(
       my_thread_init();
       error = member_actions_handler->replace_all_actions(
           exchanged_members_actions_serialized_configuration);
-      error |= set_replication_failover_channels_configuration(
-          exchanged_replication_failover_channels_serialized_configuration);
+      error |= static_cast<int>(set_replication_failover_channels_configuration(
+          exchanged_replication_failover_channels_serialized_configuration));
       my_thread_end();
     }
   }
